# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "content_url": {
            "title": "Content Url",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "generation_id": {
            "title": "Generation Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identity_scores": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity Scores"
          },
          "idx": {
            "title": "Idx",
            "type": "integer"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "generation_id",
          "idx",
          "review_status",
          "content_url",
          "created_at"
        ],
        "title": "Candidate",
        "type": "object"
      },
      "CandidateEditIn": {
        "properties": {
          "instruction": {
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          }
        },
        "required": [
          "instruction"
        ],
        "title": "CandidateEditIn",
        "type": "object"
      },
      "CandidateReviewIn": {
        "properties": {
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict"
        ],
        "title": "CandidateReviewIn",
        "type": "object"
      },
...
      "Character": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "avatar_initials": {
            "title": "Avatar Initials",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "default_style_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Style Id"
          },
          "has_canonical_ref_set": {
            "title": "Has Canonical Ref Set",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lore_md": {
            "title": "Lore Md",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "negative_traits": {
            "title": "Negative Traits",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "visual_contract": {
            "title": "Visual Contract",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "lore_md",
          "visual_contract",
          "negative_traits",
          "default_style_id",
          "created_at",
          "has_canonical_ref_set",
          "avatar_url",
          "avatar_initials"
        ],
        "title": "Character",
        "type": "object"
      },
      "CharacterInput": {
        "properties": {
          "default_style_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Style Id"
          },
          "lore_md": {
            "default": "",
            "title": "Lore Md",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "negative_traits": {
            "default": "",
            "title": "Negative Traits",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "visual_contract": {
            "default": "",
            "title": "Visual Contract",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CharacterInput",
        "type": "object"
      },
...
      "Generation": {
        "properties": {
          "actual_cost_usd_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost Usd Cents"
          },
          "aspect_ratio": {
            "title": "Aspect Ratio",
            "type": "string"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/GenerationAttachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "base_stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cost_usd_cents": {
            "title": "Cost Usd Cents",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Text"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_size": {
            "title": "Image Size",
            "type": "string"
          },
          "interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "prompt_hash": {
            "title": "Prompt Hash",
            "type": "string"
          },
          "reserved_cost_usd_cents": {
            "title": "Reserved Cost Usd Cents",
            "type": "integer"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "model",
          "image_size",
          "aspect_ratio",
          "prompt",
          "prompt_hash",
          "attachments",
          "warnings",
          "cost_usd_cents",
          "reserved_cost_usd_cents",
          "actual_cost_usd_cents",
          "state",
          "interaction_id",
          "error_text",
          "completed_at",
          "created_at",
          "candidates"
        ],
        "title": "Generation",
        "type": "object"
      },
      "GenerationAttachment": {
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "integer"
          },
          "character_name": {
            "title": "Character Name",
            "type": "string"
          },
          "image_number": {
            "title": "Image Number",
            "type": "integer"
          },
          "ref_set_id": {
            "title": "Ref Set Id",
            "type": "integer"
          },
          "ref_set_version": {
            "title": "Ref Set Version",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "image_number",
          "character_id",
          "character_name",
          "ref_set_id",
          "ref_set_version",
          "role"
        ],
        "title": "GenerationAttachment",
        "type": "object"
      },
...
      "GenerationSummary": {
        "properties": {
          "actual_cost_usd_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost Usd Cents"
          },
          "base_stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cost_usd_cents": {
            "title": "Cost Usd Cents",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Text"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "reserved_cost_usd_cents": {
            "title": "Reserved Cost Usd Cents",
            "type": "integer"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model",
          "cost_usd_cents",
          "reserved_cost_usd_cents",
          "actual_cost_usd_cents",
          "state",
          "error_text",
          "completed_at",
          "created_at",
          "candidates"
        ],
        "title": "GenerationSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LoreCraft3000",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/base-stages/{base_stage_id}/generations": {
      "get": {
        "operationId": "list_base_stage_generations_api_v1_base_stages__base_stage_id__generations_get",
        "parameters": [
          {
            "in": "path",
            "name": "base_stage_id",
            "required": true,
            "schema": {
              "title": "Base Stage Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GenerationSummary"
                  },
                  "title": "Response List Base Stage Generations Api V1 Base Stages  Base Stage Id  Generations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Base Stage Generations",
        "tags": [
          "api-v1",
          "api-v1-base-stages"
        ]
      }
    },
...
    "/api/v1/candidates/{candidate_id}/content": {
      "get": {
        "operationId": "candidate_content_api_v1_candidates__candidate_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Candidate Content",
        "tags": [
          "api-v1",
          "api-v1-generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/edit": {
      "post": {
        "description": "Create a new candidate by editing an existing one with an instruction.\n\nThe edit runs on the same provider/model that produced the source image and\nis recorded as a child generation of the source, keeping the scene's attempt\nhistory a connected chain.",
        "operationId": "edit_candidate_api_v1_candidates__candidate_id__edit_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEditIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Generation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Candidate",
        "tags": [
          "api-v1",
          "api-v1-generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/review": {
      "post": {
        "operationId": "review_candidate_api_v1_candidates__candidate_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Candidate",
        "tags": [
          "api-v1",
          "api-v1-generations"
        ]
      }
    },
    "/api/v1/characters": {
      "get": {
        "operationId": "list_characters_api_v1_characters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Character"
                  },
                  "title": "Response List Characters Api V1 Characters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Characters",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      },
      "post": {
        "operationId": "create_character_api_v1_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      }
    },
...
    "/api/v1/characters/{character_id}": {
      "delete": {
        "description": "Archive (soft-delete) a character. Existing scenes keep referencing it.",
        "operationId": "archive_character_api_v1_characters__character_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Character",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      },
      "get": {
        "operationId": "get_character_api_v1_characters__character_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      },
      "put": {
        "operationId": "update_character_api_v1_characters__character_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Character",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      }
    },
...
    "/api/v1/characters/{character_id}/restore": {
      "post": {
        "operationId": "restore_character_api_v1_characters__character_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Character",
        "tags": [
          "api-v1",
          "api-v1-characters"
        ]
      }
    },
...
    "/api/v1/generations/{generation_id}": {
      "get": {
        "operationId": "get_generation_api_v1_generations__generation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Generation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generation",
        "tags": [
          "api-v1",
          "api-v1-generations"
        ]
      }
    },
...
    "/api/v1/scenes/{scene_id}/generations": {
      "get": {
        "operationId": "list_scene_generations_api_v1_scenes__scene_id__generations_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GenerationSummary"
                  },
                  "title": "Response List Scene Generations Api V1 Scenes  Scene Id  Generations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scene Generations",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
...
  }
}